requirements.txt
pygame
matplotlib
tk
pillow